feat(views): implement products list (wip)
