+++ src/views/Products/mocks.ts
@@ -1,0 +1,13 @@
+import { Product } from '@/types'
+
+export const products: Product[] = [
+  {
+    id: '93efbb3e-beb0-4229-9a3c-d452a2f36e38',
+    name: 'Test Product',
+    vendor: {
+      id: 'f7fa92e0-4e3f-4674-b9f8-100b1c2d1bd9',
+      name: 'ABC Software Solutions',
+    },
+    lastUpload: '2023-07-25T03:47:12.798Z',
+  },
+]
