+++ src/views/Products/schema.ts
@@ -1,0 +1,14 @@
+const schema = [
+  {
+    name: 'name',
+    label: 'Name',
+    type: 'text',
+  },
+  {
+    name: 'address',
+    label: 'Address',
+    type: 'text',
+  },
+]
+
+export default schema
